move enum and stats to their own file
