+++ core/src/moe/yuru/newhorizons/models/Building.java
@@ -7,57 +7,6 @@ import com.badlogic.gdx.utils.Array;
  */
 public class Building {
 
-    /**
-     * The possible factions for a building.
-     */
-    public enum Faction {
-
-        SCIENCE, CULTURE, INDUSTRY, POLITIC
-
-    }
-
-    /**
-     * Represents the building statistics for a level. The type of the resource
-     * depends of the current faction.
-     */
-    public static class BuildingStats {
-
-        private float coinCost;
-        private float resourcesCost;
-        private float coinsPerSecond;
-        /** Raw number, the type of the resource depends of the current faction */
-        private float resourcesPerSecond;
-
-        /**
-         * @return the upgrade cost (coins)
-         */
-        public float getCoinCost() {
-            return coinCost;
-        }
-
-        /**
-         * @return the upgrade cost (resources)
-         */
-        public float getResourcesCost() {
-            return resourcesCost;
-        }
-
-        /**
-         * @return the coins generated/lost per second
-         */
-        public float getCoinsPerSecond() {
-            return coinsPerSecond;
-        }
-
-        /**
-         * @return the resources generated/lost per second
-         */
-        public float getResourcesPerSecond() {
-            return resourcesPerSecond;
-        }
-
-    }
-
     private String id;
     private String lastName;
     private String firstName;
