+++ core/src/moe/yuru/newhorizons/models/Faction.java
@@ -1,0 +1,10 @@
+package moe.yuru.newhorizons.models;
+
+/**
+ * The possible factions for a building.
+ */
+public enum Faction {
+
+    SCIENCE, CULTURE, INDUSTRY, POLITIC
+
+}
